Implement relay control, UART, and signal interpreting
Big commit.  On start, the ESP32 will configure all necessary pins and
modules (including those for UART transmission).  For each radio signal
check, the ESP32 will interpret each radio channel's value to turn
on/off the appropriate relays and send information to the motor
controllers.  In theory, this will work, although it's not been tested
yet in the robot.